fix(types): satisfy pyright strict under numpy 2.5 stubs
The lint job now resolves the project environment correctly, so pyright
checks against numpy 2.5's tightened stubs instead of whatever older
numpy an ambient environment provides.

- Replace np.sum(arr, axis=...) calls with the ndarray .sum(axis=...)
  method; the numpy 2.5 function overloads no longer resolve for these
  call shapes and degraded to Unknown
- Widen get_total_count/get_female_count/get_male_count annotations
  from int to float across BasePopulation, OutputMixin and
  AgeStructuredPopulation — the runtime values were already np.float64
  (their docstrings even said float)
- Guard nicegui slider/checkbox values (bool | None, float | None)
  in the dashboards before arithmetic and export calls
- Converge isinstance-narrowed bare ndarray types via cast in
  rust_backend and _writers

+++ src/natal/frontend/population/_mixins/_output.py
@@ -63,17 +63,17 @@ class OutputMixin(ModifierPresetMixin):
     # ── Abstract query methods (from base.py:944-957) ──────────────
 
     @abstractmethod
-    def get_total_count(self) -> int:
+    def get_total_count(self) -> float:
         """Return the total number of individuals in the population."""
         ...
 
     @abstractmethod
-    def get_female_count(self) -> int:
+    def get_female_count(self) -> float:
         """Return the total number of female individuals."""
         ...
 
     @abstractmethod
-    def get_male_count(self) -> int:
+    def get_male_count(self) -> float:
         """Return the total number of male individuals."""
         ...
 
@@ -261,17 +261,17 @@ class OutputMixin(ModifierPresetMixin):
     # From base.py:963-982
 
     @property
-    def total_population_size(self) -> int:
+    def total_population_size(self) -> float:
         """Total population size (alias of ``get_total_count``)."""
         return self.get_total_count()
 
     @property
-    def total_females(self) -> int:
+    def total_females(self) -> float:
         """Total number of females (alias of ``get_female_count``)."""
         return self.get_female_count()
 
     @property
-    def total_males(self) -> int:
+    def total_males(self) -> float:
         """Total number of males (alias of ``get_male_count``)."""
         return self.get_male_count()
